feat:restaurantBookmark

+++ server/src/main/java/application/dto/PageInfo.java
@@ -15,7 +15,7 @@ public class PageInfo {
     //페이지당 보여줄 요소의 개수 저장
     private int size;
     //전체 페이지의 개수 저장
-    private int totalPages;
+    private long totalPages;
     //전체 요소의 개수 저장
     private long totalElements;
     //전체 데이터의 개수 저장
